+++ src/input/read_time_series.jl
@@ -1,7 +1,7 @@
 """
     read_time_series(pf::PeriodsFinder, ts_name::String)
 
-Reads a time series based on data in `pf.config[ts_name]`.
+Reads the time series `ts_name` based on data provided in `pf.config[ts_name]`.
 """
 function read_time_series(pf::PeriodsFinder, ts_name::String)
 
@@ -27,7 +27,10 @@ function read_time_series(pf::PeriodsFinder, ts_name::String)
         Symbol(k) => get(csv_options_0, k, "") 
         for (k,v) in csv_options_0 if isempty(v) == false
     )
-    csv_options[:types] = Dict(val_col => Float64, timestamp_col => DateTime)
+    csv_options[:types] = Dict(val_col => Float64)
+    if isempty(timestamp_col) == false
+        csv_options[:types][timestamp_col] = DateTime
+    end
     csv_options = namedtuple(collect(csv_options))
     df = CSV.read(source, DataFrame; validate=false, csv_options...)
 
@@ -76,7 +79,7 @@ function read_time_series(pf::PeriodsFinder, ts_name::String)
     end
 
     metaDict = Dict(
-        k => v for (k, v) in merge(ts_dict[ts_name], ts_dict["default"])
+        k => v for (k, v) in merge(ts_dict["default"], ts_dict[ts_name])
     )
     metaDict["name"] = ts_name
     if haskey(metaDict, "start")
@@ -91,16 +94,5 @@ function read_time_series(pf::PeriodsFinder, ts_name::String)
     end
     metaDict["start"] = start
 
-    return ta = TimeArray(timestamps, data, [:value], metaDict)
-
-    # pf.time_series[ts_name] = readtimearray(source; kwargs...)
-
-    # push!(pf.curves, ts.name)
-
-    # weight!(pf.WEIGHT_DC, ts)
-    # area_total!(pf.AREA_TOTAL, ts)
-    # area_total_days!(pf.AREA_TOTAL_DAY, pf.periods, ts)
-
-    # cum_bin_end!(pf.A, pf.periods, pf.bins, ts)
-    # cum_bin_total!(pf.L, self.bins, ts)
+    return TimeArray(timestamps, data, [:value], metaDict)
 end