+++ src/methods/optimization.jl
@@ -66,7 +66,7 @@ end
 function make_periods_finder_model!(pf::PeriodsFinder, x::Any)
     error("""
     Second argument must be an optimizer with attributes. 
-    Perhaps you only specified `Cbc.Optimizer`, and not `optimizer_with_attributes(Cbc.Optimizer)`?
+    Perhaps you only specified e.g. `Cbc.Optimizer`, and not `optimizer_with_attributes(Cbc.Optimizer)`?
     """
     )
 end
@@ -296,7 +296,7 @@ function define_time_series_error_variable!(pf::PeriodsFinder, m::JuMP.Model)
             [s in S, i in periods, t in time_steps],
             time_series_error[s,i,t]
             == 
-            + sum(v[i,j] * x[s][i,t] for j in rep_periods)
+            + sum(v[i,j] * x[s][j,t] for j in rep_periods)
             - x[s][i,t]
         )
         @pack! m.ext[:constraints] = time_series_error
@@ -304,7 +304,7 @@ function define_time_series_error_variable!(pf::PeriodsFinder, m::JuMP.Model)
     elseif type == "absolute"
         reconstructed_time_series = @expression(m,
             [s in S, i in periods, t in time_steps],
-            sum(v[i,j] * x[s][i,t] for j in rep_periods)
+            sum(v[i,j] * x[s][j,t] for j in rep_periods)
         )
         time_series_error_eq1 = @constraint(m, 
             [s in S, i in periods, t in time_steps],
@@ -320,6 +320,7 @@ function define_time_series_error_variable!(pf::PeriodsFinder, m::JuMP.Model)
             - reconstructed_time_series[s,i,t]
             + x[s][i,t]
         )
+        @pack! m.ext[:expressions] = reconstructed_time_series
         @pack! m.ext[:constraints] = time_series_error_eq1, time_series_error_eq2 
     end
     return m