+++ test/test_days_ordering.jl
@@ -2,7 +2,7 @@ using RepresentativePeriodsFinder, JuMP, Cbc
 RPF = RepresentativePeriodsFinder
 
 # Create PeriodsFinder
-config_file = normpath(joinpath(@__DIR__, "input_data", "default.yaml"))
+config_file = RPF.datadir("default.yaml")
 pf = PeriodsFinder(config_file, populate_entries=true)
 
 # Delete entries for time series error and ordering error