+++ test/test_days_re_ordering.jl
@@ -4,7 +4,7 @@
 # First we create an instance of a PeriodsFinder. Here we are using the default `.yaml` configuration file, but keep in mind that a lot of the following lines of code could be avoided by creating your own custom file.
 using RepresentativePeriodsFinder;
 RPF = RepresentativePeriodsFinder;
-config_file = normpath(joinpath(@__DIR__, "input_data", "default.yaml"));
+config_file = RPF.datadir("default.yaml");
 pf = PeriodsFinder(config_file, populate_entries=true);
 
 # Change the result directory name.
@@ -54,9 +54,6 @@ create_synthetic_time_series_plot(pf, "Load"; timestamps=timestamps)
 
 # We can also check out the heatmap, which shows how representative days are mapped to non-representative days throughout the year:
 create_ordering_heatmap(pf)
-
-#md # !!! note
-#md #     The above heatmap should (probably) have more entries between 0 and 1, as opposed to 0 or 1. This may be due to the small number of days, the solver or be a bug.
 
 # We can also write out the this synthetic time series to `pf.config["results"]["result_dir"]`.
 # ```julia